requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
click>=8.0.0
pathspec>=0.11.0
tiktoken>=0.5.0
filelock
fsspec 
packaging 
pyyaml 
Pillow 
torch 
numpy
pandas>=1.2
python-dateutil>=2.7
joblib>=1.2.0
